requirements.txt
# deep_sort_realtime==1.3.2
# facenet_pytorch==2.6.0
# ImageHash==3.5
# joblib==1.4.2
# numpy==1.24.3
# opencv_contrib_python==4.10.0.84
# opencv_python==4.10.0.84
# opencv_python_headless==4.10.0.84
# Pillow==9.5.0
# scikit_learn==1.1.3
# supervision==0.25.1
# tensorflow==2.13.1
# torch==2.2.2
# ultralytics==8.3.55
# yolov5==7.0.14
# pipreqs==0.5.0
# https://github.com/ifzhang/ByteTrack/archive/refs/heads/master.zip
ultralytics==8.3.55
facenet-pytorch==2.6.0
joblib==1.4.2
scikit-learn==1.6.0
supervision==0.25.1
# supervision==0.13.0
opencv-contrib-python==4.10.0.84
configparser==7.1.0
filterpy==1.4.5
torch==2.2.2
torchvision==0.17.2
ImageHash==4.3.1
mediapipe==0.10.11
readchar==4.2.1
pyttsx3==2.98
gtts==2.5.4
pydub==0.25.1
mysql-connector-python== 9.0.0
# https://github.com/hanifabd/realtime-head-pose-detection/blob/master/app.py